@@ -3,7 +3,7 @@
     "license": "proprietary",
 
     "require": {
-        "php": ">=7.3",
+        "php": ">=7.4",
         "ext-SimpleXML": "*",
         "ext-gd": "*",
         "ext-imagick": "*",
@@ -13,25 +13,24 @@
         "ext-pcntl": "*",
         "ext-posix": "*",
         "ext-zip": "*",
-        "catrobat/capi": "^1.0.26",
-        "doctrine/data-fixtures": "^1.3",
-        "doctrine/doctrine-bundle": "^1.9",
-        "doctrine/doctrine-fixtures-bundle": "^3.0",
-        "doctrine/doctrine-migrations-bundle": "^2.0",
-        "doctrine/migrations": "^2.0",
-        "doctrine/orm": "^2.6",
-        "eightpoints/guzzle-bundle": "^7.3",
+        "catrobat/capi": "^1.0.27",
+        "doctrine/data-fixtures": "^1.4",
+        "doctrine/doctrine-bundle": "^2.0",
+        "doctrine/doctrine-fixtures-bundle": "^3.3",
+        "doctrine/doctrine-migrations-bundle": "^2.1",
+        "doctrine/migrations": "^2.2",
+        "doctrine/orm": "^2.7",
+        "eightpoints/guzzle-bundle": "^8.0",
         "fr3d/ldap-bundle": "^4.0",
         "friendsofsymfony/jsrouting-bundle": "^2.5",
-        "friendsofsymfony/rest-bundle": "^2.6",
+        "friendsofsymfony/rest-bundle": "^2.7",
         "friendsofsymfony/user-bundle": "^2.1",
-        "google/apiclient": "^2.2",
+        "google/apiclient": "^2.4",
         "incenteev/composer-parameter-handler": "^2.1",
-        "knplabs/knp-paginator-bundle": "^2.8",
         "lexik/jwt-authentication-bundle": "^2.6",
         "liip/theme-bundle": "dev-master",
         "php-http/guzzle6-adapter": "^2.0",
-        "php-http/message": "^1.7",
+        "php-http/message": "^1.8",
         "sensio/framework-extra-bundle": "*",
         "sonata-project/admin-bundle": "3.57",
         "sonata-project/doctrine-orm-admin-bundle": "3.13",
@@ -40,7 +39,7 @@
         "symfony/browser-kit": "*",
         "symfony/console": "*",
         "symfony/dotenv": "*",
-        "symfony/flex": "^1.2",
+        "symfony/flex": "^1.6",
         "symfony/framework-bundle": "*",
         "symfony/intl": "*",
         "symfony/maker-bundle": "*",
@@ -65,7 +64,7 @@
         "emuse/behat-html-formatter": "@dev",
         "friendsofphp/php-cs-fixer": "^2.16",
         "phpcheckstyle/phpcheckstyle": "^0.14.8",
-        "phpspec/phpspec": "^5.1",
+        "phpspec/phpspec": "^6.1",
         "phpunit/phpunit": "^8.1",
         "symfony/phpunit-bridge": "*",
         "vipsoft/doctrine-data-fixtures-extension": "3.0.*@dev",
@@ -78,14 +77,14 @@
 
     "config": {
         "platform": {
-            "php": "7.3",
-            "ext-gd": "7.3",
+            "php": "7.4",
+            "ext-gd": "7.4",
             "ext-imagick": "3.4.4",
             "ext-ldap": "2.4",
             "ext-json": "1.7.0",
-            "ext-mbstring": "7.3",
-            "ext-pcntl": "7.3",
-            "ext-posix": "7.3",
+            "ext-mbstring": "7.4",
+            "ext-pcntl": "7.4",
+            "ext-posix": "7.4",
             "ext-zip": "1.15.4"
         },
         "bin-dir": "bin",
