@@ -13,15 +13,9 @@ use Symfony\Component\DependencyInjection\ParameterBag\ParameterBagInterface;
  */
 class ArchiveLogsCommand extends Command
 {
-  /**
-   * @var
-   */
-  private $output;
+  private OutputInterface $output;
 
-  /**
-   * @var ParameterBagInterface
-   */
-  private $parameter_bag;
+  private ParameterBagInterface $parameter_bag;
 
   /**
    * ArchiveLogsCommand constructor.
@@ -39,10 +33,7 @@ class ArchiveLogsCommand extends Command
     ;
   }
 
-  /**
-   * @return int|void|null
-   */
-  protected function execute(InputInterface $input, OutputInterface $output)
+  protected function execute(InputInterface $input, OutputInterface $output): void
   {
     $this->output = $output;
 
@@ -61,8 +52,10 @@ class ArchiveLogsCommand extends Command
     {
       $affected_files .= ' '.basename($log_file);
     }
-    CommandHelper::executeShellCommand($compression_command.$affected_files,
-      ['timeout' => 7200], 'Executing command: '.$compression_command.$affected_files, $output);
+    CommandHelper::executeShellCommandLegacy(
+      $compression_command.$affected_files,
+      ['timeout' => 7200], 'Executing command: '.$compression_command.$affected_files, $output
+    );
     $this->output->writeln('Successfully archived log files');
   }
 }
