@@ -2,37 +2,29 @@
 
 namespace App\Catrobat\Commands;
 
+use App\Catrobat\Commands\Helpers\CommandHelper;
 use App\Entity\Program;
 use Doctrine\ORM\EntityManagerInterface;
+use Doctrine\ORM\NonUniqueResultException;
+use Doctrine\ORM\NoResultException;
+use Exception;
 use Symfony\Component\Console\Command\Command;
-use Symfony\Component\Console\Exception\ExceptionInterface;
-use Symfony\Component\Console\Input\ArrayInput;
 use Symfony\Component\Console\Input\InputArgument;
 use Symfony\Component\Console\Input\InputInterface;
-use Symfony\Component\Console\Output\Output;
 use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\DependencyInjection\ParameterBag\ParameterBagInterface;
-use Symfony\Component\Process\Process;
+
 
 /**
  * Class RestoreBackupCommand.
  */
 class RestoreBackupCommand extends Command
 {
-  /**
-   * @var Output
-   */
-  public $output;
+  public OutputInterface $output;
 
-  /**
-   * @var EntityManagerInterface
-   */
-  private $entity_manager;
+  private EntityManagerInterface $entity_manager;
 
-  /**
-   * @var ParameterBagInterface
-   */
-  private $parameter_bag;
+  private ParameterBagInterface $parameter_bag;
 
   /**
    * RestoreBackupCommand constructor.
@@ -54,12 +46,11 @@ class RestoreBackupCommand extends Command
   }
 
   /**
-   * @throws ExceptionInterface
-   * @throws \Doctrine\ORM\NonUniqueResultException | \Exception
-   *
-   * @return int|void|null
+   * @throws NoResultException
+   * @throws NonUniqueResultException
+   * @throws Exception
    */
-  protected function execute(InputInterface $input, OutputInterface $output)
+  protected function execute(InputInterface $input, OutputInterface $output): void
   {
     $this->output = $output;
 
@@ -70,14 +61,14 @@ class RestoreBackupCommand extends Command
 
       if (!is_file($backup_file))
       {
-        throw new \Exception('File not found');
+        throw new Exception('File not found');
       }
     }
     $this->output->writeln('Backup File: '.$backup_file);
 
     if ('pdo_mysql' !== $_ENV['DATABASE_DRIVER'])
     {
-      throw new \Exception('This script only supports mysql databases');
+      throw new Exception('This script only supports mysql databases');
     }
 
     if ($input->hasArgument('local') && 'true' === $input->getArgument('local'))
@@ -89,26 +80,29 @@ class RestoreBackupCommand extends Command
 
       $this->output->writeln('Restore backup on localhost');
 
-      $this->executeSymfonyCommand('catrobat:purge', ['--force' => true], $this->output);
-
-      $this->executeShellCommand("time pigz -dc {$backup_file} | tar -xf - -C {$local_resource_directory}",
-        'Extracting files to local resources directory');
-
-      $this->executeShellCommand("time mysql -u {$local_database_user} -p{$local_database_password} {$local_database_name} ".
-        '< '.$local_resource_directory.'database.sql ',
-        'Restore SQL file');
+      CommandHelper::executeSymfonyCommand(
+        'catrobat:purge', $this->getApplication(), ['--force' => true], $this->output
+      );
+      CommandHelper::executeShellCommandLegacy(
+        "time pigz -dc {$backup_file} | tar -xf - -C {$local_resource_directory}", ['timeout' => 86400],
+        'Extracting files to local resources directory', $output
+      );
+      CommandHelper::executeShellCommandLegacy(
+        "time mysql -u {$local_database_user} -p{$local_database_password} {$local_database_name} ".
+        '< '.$local_resource_directory.'database.sql ', ['timeout' => 86400], 'Restore SQL file', $output
+      );
 
       @unlink($local_resource_directory.'database.sql');
 
       $query = $this->entity_manager
-        ->createQuery('UPDATE App\\Entity\\Program p SET p.apk_status = :status WHERE p.apk_status != :status')
+        ->createQuery('UPDATE App\Entity\Program p SET p.apk_status = :status WHERE p.apk_status != :status')
       ;
       $query->setParameter('status', Program::APK_NONE);
       $result = $query->getSingleScalarResult();
       $this->output->writeln('Reset the apk status of '.$result.' projects');
 
       $query = $this->entity_manager
-        ->createQuery('UPDATE App\\Entity\\Program p SET p.directory_hash = :hash WHERE p.directory_hash != :hash')
+        ->createQuery('UPDATE App\Entity\Program p SET p.directory_hash = :hash WHERE p.directory_hash != :hash')
       ;
       $query->setParameter('hash', 'null');
       $result = $query->getSingleScalarResult();
@@ -128,74 +122,42 @@ class RestoreBackupCommand extends Command
 
       $this->output->writeln('Restore backup on server ['.$backup_host_name.']');
 
-      $this->executeShellCommand("sshpass -p '{$backup_host_password}' ssh {$backup_host_user}@{$backup_host_name} ".
-        '"php '.$backup_host_directory.'bin/console catrobat:purge --env=prod --force" ',
-        'Remove files from server resources directory');
-
-      $this->executeShellCommand("gzip -dc {$backup_file} | ".
+      CommandHelper::executeShellCommandLegacy(
         "sshpass -p '{$backup_host_password}' ssh {$backup_host_user}@{$backup_host_name} ".
-        'tar -xf - -C '.$backup_host_directory.$backup_host_resource_directory.' ',
-        'Extract files to server resources directory');
-
-      $this->executeShellCommand("sshpass -p '{$backup_host_password}' ssh {$backup_host_user}@{$backup_host_name} ".
+        '"php '.$backup_host_directory.'bin/console catrobat:purge --env=prod --force" ', ['timeout' => 86400],
+        'Remove files from server resources directory', $output
+      );
+      CommandHelper::executeShellCommandLegacy(
+        "gzip -dc {$backup_file} | ".
+        "sshpass -p '{$backup_host_password}' ssh {$backup_host_user}@{$backup_host_name} ".
+        'tar -xf - -C '.$backup_host_directory.$backup_host_resource_directory.' ', ['timeout' => 86400],
+        'Extract files to server resources directory', $output
+      );
+      CommandHelper::executeShellCommandLegacy(
+        "sshpass -p '{$backup_host_password}' ssh {$backup_host_user}@{$backup_host_name} ".
         "\"mysql -u {$backup_database_user} -p{$backup_database_password} {$backup_database_name} ".
-        '< '.$backup_host_directory.$backup_host_resource_directory.'database.sql" ',
-        'Restore SQL file');
-
-      $this->executeShellCommand("sshpass -p '{$backup_host_password}' ssh {$backup_host_user}@{$backup_host_name} ".
-        '"unlink '.$backup_host_directory.$backup_host_resource_directory.'database.sql" ',
-        'Remove files from server resources directory');
-
-      $this->executeShellCommand("sshpass -p '{$backup_host_password}' ssh {$backup_host_user}@{$backup_host_name} ".
+        '< '.$backup_host_directory.$backup_host_resource_directory.'database.sql" ', ['timeout' => 86400],
+        'Restore SQL file', $output
+      );
+      CommandHelper::executeShellCommandLegacy(
+        "sshpass -p '{$backup_host_password}' ssh {$backup_host_user}@{$backup_host_name} ".
+        '"unlink '.$backup_host_directory.$backup_host_resource_directory.'database.sql" ', ['timeout' => 86400],
+        'Remove files from server resources directory', $output
+      );
+      CommandHelper::executeShellCommandLegacy(
+        "sshpass -p '{$backup_host_password}' ssh {$backup_host_user}@{$backup_host_name} ".
         "\"mysql -u {$backup_database_user} -p{$backup_database_password} {$backup_database_name} ".
         "-e 'UPDATE program p SET p.apk_status = ".Program::APK_NONE.' WHERE p.apk_status != '.Program::APK_NONE."'\" ",
-        'Reset the apk status');
-
-      $this->executeShellCommand("sshpass -p '{$backup_host_password}' ssh {$backup_host_user}@{$backup_host_name} ".
+        ['timeout' => 86400], 'Reset the apk status', $output
+      );
+      CommandHelper::executeShellCommandLegacy(
+        "sshpass -p '{$backup_host_password}' ssh {$backup_host_user}@{$backup_host_name} ".
         "\"mysql -u {$backup_database_user} -p{$backup_database_password} {$backup_database_name} ".
         "-e 'UPDATE program p SET p.directory_hash = \"null\" WHERE p.directory_hash != \"null\"'\" ",
-        'Reset the directory hash');
+        ['timeout' => 86400],'Reset the directory hash', $output
+      );
     }
 
     $this->output->writeln('Import finished!');
-  }
-
-  /**
-   * @param $command
-   * @param $description
-   *
-   * @throws \Exception
-   *
-   * @return bool
-   */
-  private function executeShellCommand($command, $description)
-  {
-    $this->output->write($description." ('".$command."') ... ");
-    $process = new Process($command);
-    $process->setTimeout(86400);
-    $process->run();
-    if ($process->isSuccessful())
-    {
-      $this->output->writeln('OK');
-
-      return true;
-    }
-
-    throw new \Exception('failed: '.$process->getErrorOutput());
-  }
-
-  /**
-   * @param $command
-   * @param $args
-   * @param $output
-   *
-   * @throws ExceptionInterface
-   */
-  private function executeSymfonyCommand($command, $args, $output)
-  {
-    $command = $this->getApplication()->find($command);
-    $args['command'] = $command;
-    $input = new ArrayInput($args);
-    $command->run($input, $output);
   }
 }
